User class

+++ python/user/decimalencoder.py
@@ -5,5 +5,5 @@ import json
 class DecimalEncoder(json.JSONEncoder):
     def default(self, obj):
         if isinstance(obj, decimal.Decimal):
-            return int(obj)
+            return float(obj)
         return super(DecimalEncoder, self).default(obj)
